// src/frontend/src/backend.d.ts
export interface SavedItem {
    rate: bigint;
    description: string;
    itemCode: string;
    itemName: string;
    taxRate: bigint;
    taxType: Variant_igst_none_cgst_sgst;
}
// ...
export enum Variant_igst_none_cgst_sgst {
    igst = "igst",
    none = "none",
    cgst_sgst = "cgst_sgst"
}